tests cleaned, split files, ready to PR
