+++ src/action-types.js
@@ -1,27 +1,4 @@
-import { TypeError, UnicityError } from './errors.js';
-
-const isString = arg => typeof arg === 'string' || arg instanceof String;
-
-const StringOrThrow = element => {
-  if (!isString(element)) throw new Error(TypeError);
-};
-
-const RaiseErrorIfNotUnique = array => {
-  if (array.length !== new Set(array).size) {
-    throw new Error(UnicityError);
-  }
-};
-
-const actionTypesWithArray = (namespace, constants) => {
-  RaiseErrorIfNotUnique(constants);
-  return Object.freeze(
-    constants.reduce((obj, constant) => {
-      StringOrThrow(constant);
-      obj[constant] = `${namespace}/${constant}`;
-      return obj;
-    }, {}),
-  );
-};
+import actionTypesWithArray from './action-types-array';
 
 export default function actionTypes(namespace, ...constants) {
   console.warn(
